style: rustfmt and clippy across the tree, and what is left is not style
Mechanical only, in its own commit so it can be read as "this is rustfmt" and
skipped, or dropped, without touching the change it sits beside.

    rustfmt   496 diffs -> 0   (101 files)
    clippy     50 errors -> 13 (cargo clippy --fix)

**The remaining 13 are not lint noise and are deliberately NOT fixed here.** All
of them are in `misaka-palw-base0-ref2`, and every one is that crate's OWN
`#![deny(clippy::arithmetic_side_effects)]` firing on its own code — a policy it
declared and has never satisfied, which nobody could see because the Lints job
died in an unbuildable crate long before clippy ran.

Nine are SHA-256 index arithmetic (`w[i - 15]` inside `for i in 16..64`) and one
is `(message.len() as u64) * 8`; those are provable and want a justification, not
a rewrite. The other three are the reason this stops here:

    fn trunc_div(numerator: i128, divisor: i128) -> i128 { numerator / divisor }
    fn round_half_away(..) { .. if numerator < 0 { -rounded } else { rounded } }

`divisor == 0` and `i128::MIN / -1` panic; `-i128::MIN` panics. `floor_div`'s
doc says "divisor must be positive" and nothing checks it, while `pow2` in the
same file states its precondition with `.expect("callers bound n at 62")` — so
the crate's own convention is to assert, and these three do not.

This is an ADJUDICATION reference. A court that cannot return a verdict because
its reference panicked is a liveness failure, not a lint. Choosing between
asserting the precondition, saturating, or proving `i128::MIN` unreachable is a
decision about what the reference SPECIFIES at its boundary, and it does not
belong in a commit whose subject is formatting — nor in a branch about bond
registration.

185 suites, 2763 tests, 0 failures with these changes applied.

+++ misaka-palw-reference2/tests/differential_v2.rs
@@ -19,8 +19,7 @@ use kaspa_consensus_core::palw_reference::{
 };
 use misaka_palw_reference2::{
     REF2_CANONICAL_NAN16, REF2_CANONICAL_NAN64, ref2_add64, ref2_div, ref2_div64, ref2_f16_to_f32, ref2_f32_to_f16, ref2_fma,
-    ref2_fma64, ref2_mul64, ref2_narrow_f64_to_f32, ref2_neg64, ref2_sqrt, ref2_sub64, ref2_sub64_direct,
-    ref2_widen_f32_to_f64,
+    ref2_fma64, ref2_mul64, ref2_narrow_f64_to_f32, ref2_neg64, ref2_sqrt, ref2_sub64, ref2_sub64_direct, ref2_widen_f32_to_f64,
 };
 
 const SIGN_MASK: u32 = 0x8000_0000;
@@ -427,7 +426,7 @@ fn f32_to_f16_random_500k_agree_exactly() {
         assert_eq!(ref2_f32_to_f16(a), ref_f32_to_f16_v2(a), "CRITICAL f32->f16 disagreement #{i} a={a:08x}");
         // And in the f16-representable neighborhood, where rounding actually decides:
         // squash the exponent into [104, 143] (2^-24 .. 2^16).
-        let squashed = (a & 0x807F_FFFF) | (((104 + (rng.next_u32() % 40)) as u32) << 23);
+        let squashed = (a & 0x807F_FFFF) | ((104 + (rng.next_u32() % 40)) << 23);
         assert_eq!(
             ref2_f32_to_f16(squashed),
             ref_f32_to_f16_v2(squashed),
